requirements.txt
openai
convokit
pyyaml
python-dotenv
tqdm
